Debugging 6

+++ Task 5/Debugging6/Debugging6.pde
@@ -2,20 +2,19 @@ boolean jobsDone = true;
 
 void setup()
 {
-    
-    println(getRandomNumber(0, 10));
-    if (isJobDone())
-    {
-        println("Job's done!");   
-    }
+  println(getRandomNumber(0, 10));
+  if (isJobDone()) {
+    println("Job's done!");   
+  }
 }
 
 boolean isJobDone()
 {
-    return jobsDone;    
+  return jobsDone;    
 }
 
 int getRandomNumber(int min, int max) 
 {
-    return random(min, max);
+  // Changed returntype from method random() from float to int with typecast
+  return (int) random(min, max);
 }
